Create load run.bas
# LoadAndRun---C64

Here is a utility program for everyone. One of the worst features of the C64 
is the way the 1541 handles machine language program. It is essential that 
you load a ML program wit a “,8,1” suffix so that the computer knows where to 
locate the program in its memory. This program coverts ML programs into a BASIC 
load and run format. You never need to remember another SYS command (except to 
reactivate some routines).

HINT: Once in BASIC form you don’t need file copiers to put programs on other 
disks, simple LOAD it and then SAVE it.